+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fixed QR code background defaulting to transparent instead of white (regression in 0.4.1)
+- Fixed section margins, added support for `pageMargins: 'inherit'`
 
 ## v0.4.1 - 2026-07-13